refactor: 인터페이스는 모델의 일부이기 떄문에 적절한 패키지로 이동
