+++ src/main/java/model/lottonumber/LottoNumbersGenerationStrategy.java
@@ -1,8 +1,6 @@
-package strategy;
+package model.lottonumber;
 
 import java.util.List;
-
-import model.lottonumber.LottoNumber;
 
 public interface LottoNumbersGenerationStrategy {
 	List<LottoNumber> generate(int size);
